Added Macros
Macros allow a single ID/Barcode to be used to represent a set of items. These items can be set and updated through the admin interface.
An additional Table needs to be created in the DB to support macros, check the docs (next commit) for the SQL code.

Docs and Unit Tests will be included in future commits.